+++ packages/uploader-server/src/constants.ts
@@ -1,0 +1,1 @@
+export const ImageBinDid = 'z8ia1mAXo8ZE7ytGF36L5uBf9kD2kenhqFGp9';